Emission factor for buses in Brazil

Returns a vector or data.table of emission factors for buses based on estimates from the Environment Company of Sao Paulo, Brazil (CETESB) 2019, and obtained via vein package. Estimates are expressed in units 'g/km'.

ef_brazil_cetesb(pollutant, veh_type, model_year, as_list = TRUE)

Source

https://cetesb.sp.gov.br/veicular/relatorios-e-publicacoes/

Arguments

pollutant

character; Pollutants "CH4", "CO2", "PM10", "N2O", "NOx", "NO2", "NO", "RCHO", "ETOH" "KML", "FC" (Fuel Consumption), "gD/KWH" (grams of Diesel per kWh), "gCO2/KWH" (grams of CO2 per per kWh), "CO", "HC" (Total Hydrocarbon), "NMHC" (Non-Methane Hydrocarbon), "FS"(Fuel Sales) and "NH3".

veh_type

character; Vehicle categories by fuel: "BUS_URBAN_D" , "BUS_MICRO_D", "BUS_COACH_D" and "BUS_ARTIC_D".

model_year

numeric; Vehicle model year. Supports model_year from 1960 to 2020.

as_list

logical; Returns emission factors as a list, instead of data.table format. Default is TRUE.

Value

data.table; Emission factors in units 'g/km' by speed and model_year.

Details

The new convention for vehicles names are translated from CETESB report:

vehicle	description

BUS_URBAN_D	Urban Bus Diesel (5perc bio-diesel)

BUS_MICRO_D	Micro Urban Bus Diesel (5perc bio-diesel)

BUS_COACH_D	Coach (inter-state) Bus Diesel (5perc bio-diesel)

BUS_ARTIC_D	Articulated Urban Bus Diesel (5perc bio-diesel)

The percentage varies of biofuels varies by law.

These emission factors are not exactly the same as the report of CETESB.

In this emission factors, there is also NO and NO2 based on split by published in the EMEP/EEA air pollutant emission inventory guidebook.

Also, the emission factors were extended till 50 years of use, repeating the oldest value.

See also

Examples

if (interactive()) {
ef_brazil_cetesb(pollutant = c("CO","PM10","CO2","CH4","NOx"),veh_type = "BUS_URBAN_D",
model_year = 2015,as_list = TRUE)
}